+++ domain/src/main/java/camp/nextstep/edu/domain/MemosRepository.kt
@@ -1,0 +1,17 @@
+package camp.nextstep.edu.domain
+
+/**
+ * Created By Malibin
+ * on 8월 24, 2021
+ */
+
+interface MemosRepository {
+
+    fun save(memo: Memo)
+
+    fun getAllMemos(): List<Memo>
+
+    fun getMemo(id: String): Memo?
+
+    fun deleteMemo(id: String)
+}
